feat: crate a new module folder for api results
